Add files via upload
You are given a string s. Each character is either 0 or 1.

You want all 1s in the string to form a contiguous subsegment. For example, if the string is 0, 1, 00111 or 01111100, then all ones (1) form a contiguous subsegment, and if the string is 0101, 100001 or 11111111111101, then this condition is not met.

You may erase some (possibly none) zeroes(0) from the string. What is the minimum number of zeros (0) that you have to erase?

INPUT:
The first line contains one integer t (1≤t≤100) — the number of test cases.

Then t lines follow, each representing a test case. Each line contains one string s (1≤|s|≤100); each character of s is either 0 or 1.

OUTPUT:
Print t integers, where the i-th integer is the answer to the i-th testcase (the minimum number of zeros(0) that you have to erase from s).

Example
Input
3

010011

0

1111000

Output
2

0

0

Note
In the first test case you have to delete the third and forth symbols from string 010011 (it turns into 0111).